EmailAPI changes commit

+++ src/main/java/com/amitit/webapp/entity/Course.java
@@ -22,7 +22,7 @@ public class Course {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long cid;
+	private int cid;
 	
 	private String name;
 	private String description;
